Move mark occurrences logic out of ScalaSourceFileEditor
The code is copied to MarkOccurrencesEditorExtension. The only changes
are:

- `getSourceViewer` is replaced by `sourceViewer`
- `isActiveEditor` is replaced by `EditorUtils.isActiveEditor(this)`
- `occurrencesFinder` is now a lazy val
